+++ backend/PsyTest/PsyTest/Test.cpp
@@ -37,7 +37,7 @@ PsyTest::Test::Test(const Test & obj)
 	if (this->img_size > 0)
 	{
 		this->img = new char[this->img_size];
-		for (int i = 0; i < this->img_size; ++i)
+		for (unsigned long i = 0; i < this->img_size; ++i)
 		{
 			this->img[i] = obj.img[i];
 		}
